% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{badge_386}
\alias{badge_386}
\title{Create a Bootstrap 386 badge}
\usage{
badge_386(..., status, rounded = FALSE)
}
\arguments{
\item{...}{Text.}

\item{status}{Badge status.}

\item{rounded}{Rounded style. Default to FALSE.}
}
\value{
A shiny tags
}
\description{
Create a Bootstrap 386 badge
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)
 ui <- page_386(
  badge_386(status = "danger", "1"),
  badge_386(status = "info", "2"),
  badge_386(status = "success", "3", rounded = TRUE)
 )

 server <- function(input, output, session) {}
 shinyApp(ui, server)
}
}
