% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{button_386}
\alias{button_386}
\title{Bootstrap 386 action button}
\usage{
button_386(inputId, label, status = NULL, icon = NULL, width = NULL, ...)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{status}{Button color.}

\item{icon}{An optional \code{\link[shiny:icon]{icon()}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
A shiny tag.
}
\description{
Bootstrap 386 action button
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  button_386(
   "btn",
   HTML(paste("Value", textOutput("val"), sep = ":")),
   icon = icon("thumbs-up"),
   class = "btn-lg"
  )
 )

 server <- function(input, output) {
   output$val <- renderText(input$btn)
 }

 shinyApp(ui, server)
}
}
