% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{card_386}
\alias{card_386}
\alias{card_title_386}
\alias{card_subtitle_386}
\alias{card_link_386}
\title{Create a Bootstrap 386 card}
\usage{
card_386(..., title = NULL, status = NULL, footer = NULL)

card_title_386(title)

card_subtitle_386(subtitle)

card_link_386(href, label)
}
\arguments{
\item{...}{Card content.}

\item{title}{Title text.}

\item{status}{Card background status.}

\item{footer}{Card footer.}

\item{subtitle}{Card subtitle.}

\item{href}{Target url.}

\item{label}{Link text.}
}
\value{
A shiny tag

A shiny tag.

A shiny tag.

A shiny tag.
}
\description{
Create a Bootstrap 386 card

Create a Bootstrap 4 card title element

Create a Bootstrap 386 card subtitle element

Create a Bootstrap 386 card link element
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)
 ui <- page_386(
  card_386(
   title = "My card",
   "This is my card",
   br(),
   card_link_386(href = "https://www.google.com", "More"),
   footer = "Card footer"
  )
 )

 server <- function(input, output, session) {}
 shinyApp(ui, server)
}
}
