% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{checkbox_group_input_386}
\alias{checkbox_group_input_386}
\title{Create a Bootstrap 386 checkbox group input}
\usage{
checkbox_group_input_386(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  width = NULL,
  choiceNames = NULL,
  choiceValues = NULL
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{choices}{List of values to show checkboxes for. If elements of the list
are named then that name rather than the value is displayed to the user. If
this argument is provided, then \code{choiceNames} and \code{choiceValues}
must not be provided, and vice-versa. The values should be strings; other
types (such as logicals and numbers) will be coerced to strings.}

\item{selected}{The values that should be initially selected, if any.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{choiceNames, choiceValues}{List of names and values, respectively,
that are displayed to the user in the app and correspond to the each
choice (for this reason, \code{choiceNames} and \code{choiceValues}
must have the same length). If either of these arguments is
provided, then the other \emph{must} be provided and \code{choices}
\emph{must not} be provided. The advantage of using both of these over
a named list for \code{choices} is that \code{choiceNames} allows any
type of UI object to be passed through (tag objects, icons, HTML code,
...), instead of just simple text. See Examples.}
}
\description{
Create a Bootstrap 386 checkbox group input
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  checkbox_group_input_386("variable", "Variables to show:",
   c("Cylinders" = "cyl",
     "Transmission" = "am",
     "Gears" = "gear")),
  tableOutput("data")
 )

 server <- function(input, output, session) {
   output$data <- renderTable({
    mtcars[, c("mpg", input$variable), drop = FALSE]
   }, rownames = TRUE)
 }
 shinyApp(ui, server)

}
}
