% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{checkbox_input_386}
\alias{checkbox_input_386}
\title{Create a Bootstrap 386 checkbox}
\usage{
checkbox_input_386(inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}
}
\description{
Create a Bootstrap 386 checkbox
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  checkbox_input_386("check", "Check me", TRUE),
  verbatimTextOutput("val")
 )

 server <- function(input, output, session) {
   output$val <- renderPrint(input$check)
 }
 shinyApp(ui, server)
}
}
