% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-shiny.R
\name{create_checkbox_tag}
\alias{create_checkbox_tag}
\title{Create checkbox/switch input based on the selected type}
\usage{
create_checkbox_tag(
  inputId,
  label,
  value = FALSE,
  width = NULL,
  type = c("switch", "checkbox")
)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}

\item{type}{Input type. This is to be able to distinguish between switch and checkbox,
which have slightly different design.}
}
\value{
An input tag.
}
\description{
Used internally by \link{toggle_input_386} and \link{checkbox_input_386}
}
