% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{jumbotron_386}
\alias{jumbotron_386}
\title{Create a Bootstrap 386 Jumbotron}
\usage{
jumbotron_386(..., title = NULL)
}
\arguments{
\item{...}{Any element.}

\item{title}{Jumbotron title.}
}
\value{
A shiny tag
}
\description{
Create a Bootstrap 386 Jumbotron
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  jumbotron_386(
   title = "Jumbotron 386",
   p("Hello World"),
   button_386("btn", "More", class = "btn-lg")
  )
 )

 server <- function(input, output, session) {}

 shinyApp(ui, server)
}
}
