% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{navbar_page_386}
\alias{navbar_page_386}
\alias{navbar_menu_386}
\title{Create a Bootstrap 386 navbar page}
\usage{
navbar_page_386(
  title,
  ...,
  id = NULL,
  selected = NULL,
  position = c("static-top", "fixed-top", "fixed-bottom"),
  header = NULL,
  footer = NULL,
  inverse = FALSE,
  windowTitle = title
)

navbar_menu_386(title, ..., menuName = title, icon = NULL)
}
\arguments{
\item{title}{The title to display in the navbar}

\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the page. The
\code{navbarMenu} function also accepts strings, which will be used as menu
section headers. If the string is a set of dashes like \code{"----"} a
horizontal separator will be displayed in the menu.}

\item{id}{If provided, you can use \code{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{position}{Determines whether the navbar should be displayed at the top
of the page with normal scrolling behavior (\code{"static-top"}), pinned at
the top (\code{"fixed-top"}), or pinned at the bottom
(\code{"fixed-bottom"}). Note that using \code{"fixed-top"} or
\code{"fixed-bottom"} will cause the navbar to overlay your body content,
unless you add padding, e.g.: \code{tags$style(type="text/css", "body
  {padding-top: 70px;}")}}

\item{header}{Tag or list of tags to display as a common header above all
tabPanels.}

\item{footer}{Tag or list of tags to display as a common footer below all
tabPanels}

\item{inverse}{\code{TRUE} to use a dark background and light text for the
navigation bar}

\item{windowTitle}{The title that should be displayed by the browser window.
Useful if \code{title} is not a string.}

\item{menuName}{A name that identifies this \code{navbarMenu}. This
is needed if you want to insert/remove or show/hide an entire
\code{navbarMenu}.}

\item{icon}{Optional icon to appear on a \code{navbarMenu} tab.}
}
\value{
A shiny tag
}
\description{
Create a Bootstrap 386 navbar page

Create a Bootstrap 386 navbar menu
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- navbar_page_386(
  "App Title",
  id = "tabset",
  tab_panel_386(
    "Tab 1",
    radio_input_386(
      "dist", "Distribution type:",
      c("Normal" = "norm",
        "Uniform" = "unif",
        "Log-normal" = "lnorm",
        "Exponential" = "exp")
    ),
    plotOutput("distPlot")
  ),
  tab_panel_386(
    "Tab 2",
    select_input_386(
      "variable", "Variable:",
      c("Cylinders" = "cyl",
        "Transmission" = "am",
        "Gears" = "gear")
    ),
    tableOutput("data")
  ),
  navbar_menu_386(
   "More",
   tab_panel_386("Summary", "Extra content 1"),
   "----",
   "Section header",
   tab_panel_386("Table", "Extra content 2")
  )
 )

 server <- function(input, output, session) {
   output$distPlot <- renderPlot({
     dist <- switch(input$dist,
                    norm = rnorm,
                    unif = runif,
                    lnorm = rlnorm,
                    exp = rexp,
                    rnorm)

     hist(dist(500))
   })

   output$data <- renderTable({
     mtcars[, c("mpg", input$variable), drop = FALSE]
   }, rownames = TRUE)

   observe(print(input$tabset))

 }
 shinyApp(ui, server)
}
}
