% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{page_386}
\alias{page_386}
\title{Create Bootstrap 386 page skeleton}
\usage{
page_386(...)
}
\arguments{
\item{...}{Slot for shiny386 layout elements.}
}
\value{
A list of tags
}
\description{
Create Bootstrap 386 page skeleton
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)
}
}
