% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{progress_386}
\alias{progress_386}
\title{Create a Bootstrap 386 progress bar}
\usage{
progress_386(id = NULL, value, status = NULL)
}
\arguments{
\item{id}{Progress unique id.}

\item{value}{Progress value. Numeric between 0 and 100.}

\item{status}{Progress status.}
}
\value{
A progress bar tag.
}
\description{
The progress bar may be updated server side. See \link{update_progress_386}.
}
\seealso{
\link{update_progress_386}
}
