% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user-feedback.R
\name{show_toast_386}
\alias{show_toast_386}
\title{Show a Bootstrap 386 toast on the client}
\usage{
show_toast_386(id, options = NULL, session = getDefaultReactiveDomain())
}
\arguments{
\item{id}{Toast id.}

\item{options}{Toast options: see \url{https://getbootstrap.com/docs/4.3/components/toasts/}.}

\item{session}{Shiny session}
}
\description{
Show a Bootstrap 386 toast on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  toast_386(
    id = "toast",
    title = "Hello",
    subtitle = "now",
    "Toast body",
    img = "https://preview-dev.tabler.io/static/logo.svg"
  ),
  button_386("launch", "Go!", class = "btn-lg")
 )

 server <- function(input, output, session) {
   observe(print(input$toast))
   observeEvent(input$launch, {
     removeNotification("notif")
     show_toast_386(
       "toast",
       options = list(
         animation = FALSE,
         delay = 3000
       )
     )
   })

   observeEvent(input$toast, {
     showNotification(
       id = "notif",
       "Toast was closed",
       type = "warning",
       duration = 1,

     )
   })
 }

 shinyApp(ui, server)
}
}
