% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{tabset_panel_386}
\alias{tabset_panel_386}
\title{Create a Bootstrap 386 tabset panel}
\usage{
tabset_panel_386(
  ...,
  id = NULL,
  selected = NULL,
  type = c("tabs", "pills"),
  position = NULL
)
}
\arguments{
\item{...}{\code{\link[shiny:tabPanel]{tabPanel()}} elements to include in the tabset}

\item{id}{If provided, you can use \verb{input$}\emph{\code{id}} in your
server logic to determine which of the current tabs is active. The value
will correspond to the \code{value} argument that is passed to
\code{\link[shiny:tabPanel]{tabPanel()}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first
tab will be selected.}

\item{type}{\describe{
\item{\code{"tabs"}}{Standard tab look}
\item{\code{"pills"}}{Selected tabs use the background fill color}
\item{\code{"hidden"}}{Hides the selectable tabs. Use \code{type = "hidden"} in
conjunction with \code{\link[shiny:tabPanelBody]{tabPanelBody()}} and \code{\link[shiny:updateTabsetPanel]{updateTabsetPanel()}} to control the
active tab via other input controls. (See example below)}
}}

\item{position}{Tabs position (left or right).}
}
\description{
Create a Bootstrap 386 tabset panel
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  tabset_panel_386(
    id = "tabset",
    selected = "Tab 2",
    tabPanel("Tab 1", "Content 1"),
    tabPanel("Tab 2", "Content 2")
  )
 )

 server <- function(input, output, session) {
   observe(print(input$tabset))
 }
 shinyApp(ui, server)

}
}
