% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{toast_386}
\alias{toast_386}
\title{Create a Bootstrap 386 toast}
\usage{
toast_386(id, title = NULL, subtitle = NULL, ..., img = NULL)
}
\arguments{
\item{id}{Unique toast id.}

\item{title}{Toast title.}

\item{subtitle}{Toast subtitle.}

\item{...}{Toast content.}

\item{img}{Toast image.}
}
\value{
A toast
}
\description{
Display user feedback
}
\seealso{
\link{show_toast_386}
}
