% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{toggle_input_386}
\alias{toggle_input_386}
\title{Custom Bootstrap 386 switch input}
\usage{
toggle_input_386(inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny:validateCssUnit]{validateCssUnit()}}.}
}
\value{
A toggle input tag.
}
\description{
Similar to the shiny checkbox
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)

 ui <- page_386(
  toggle_input_386("toggle", "Toggle me", TRUE),
  verbatimTextOutput("val")
 )

 server <- function(input, output, session) {
   output$val <- renderPrint(input$toggle)
 }
 shinyApp(ui, server)
}
}
\seealso{
\link{update_toggle_input_386}.
}
