% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{update_progress_386}
\alias{update_progress_386}
\title{Update a \link{progress_386} on the client}
\usage{
update_progress_386(id, value, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{Progress unique id.}

\item{value}{New value.}

\item{session}{Shiny session object.}
}
\description{
Update a \link{progress_386} on the client
}
\examples{
if (interactive()) {
 library(shiny)
 library(shiny386)
 ui <- page_386(
  button_386("update", "Update"),
  br(),
  progress_386(id = "progress1", 12)
 )

 server <- function(input, output, session) {
   observeEvent(input$update, {
     update_progress_386(
       id = "progress1",
       sample(1:100, 1)
     )
   })
 }
 shinyApp(ui, server)
}
}
