% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{update_text_input_386}
\alias{update_text_input_386}
\title{Update a Bootstrap 386 text input on the client}
\usage{
update_text_input_386(
  session = getDefaultReactiveDomain(),
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}. Default is \code{getDefaultReactiveDomain()}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to what can
be entered into the control. Internet Explorer 8 and 9 do not support this
option.}
}
\description{
Update a Bootstrap 386 text input on the client
}
