% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{use_bs4_deps}
\alias{use_bs4_deps}
\title{Create shiny386 dependencies}
\usage{
use_bs4_deps(tag)
}
\arguments{
\item{tag}{Tag on which to add dependencies. We usually target the body.}
}
\description{
Add all necessary dependencies so that shiny386 renders well
}
\seealso{
\link{page_386}.
}
