% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_progress_value}
\alias{validate_progress_value}
\title{Validate a \link{progress_386} value}
\usage{
validate_progress_value(value)
}
\arguments{
\item{value}{Value to validate.}
}
\value{
An error is raised if the value does not met expectations.
}
\description{
Validate a \link{progress_386} value
}
