% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{update_text_input_386}
\alias{update_text_input_386}
\title{Update a Bootstrap 386 text input on the client}
\usage{
update_text_input_386(
  session,
  inputId,
  label = NULL,
  value = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{The \code{session} object passed to function given to
\code{shinyServer}.}

\item{inputId}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{The value to set for the input object.}

\item{placeholder}{The placeholder to set for the input object.}
}
\description{
Update a Bootstrap 386 text input on the client
}
